pyproject.toml
[tool.poetry]
name = "social_network"
version = "0.1.0"
description = ""
authors = ["vladimir <webmagic@ukr.net>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
django-environ = "^0.4.5"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
django-extensions = "^3.1.2"
Pillow = "^8.2.0"
ipython = "^7.22.0"
djangorestframework-simplejwt = "^4.6.0"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
drf-yasg = "^1.20.0"
argon2-cffi = "^20.1.0"
factory-boy = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
